Move entities and facsimiles
Entities and facsimiles where present on the layer, but now they are
moved to a lookup in entry.textData. On the layers, only the IDs of the
entities and facsimiles are present. Also the active facsimile and
entities are moved to entry.activeFacsimiles and entry.activeEntities.
This way the entry does not have to change and only the active
facsimiles, active entities and layers are stateful.